+++ src/main/java/ml/pluto7073/bartending/foundations/item/PourableBottleItem.java
@@ -4,6 +4,7 @@ import ml.pluto7073.bartending.content.sound.BartendingSounds;
 import ml.pluto7073.bartending.foundations.BartendingStats;
 import ml.pluto7073.bartending.foundations.recipe.BartendingRecipes;
 import ml.pluto7073.bartending.foundations.recipe.PouringRecipe;
+import ml.pluto7073.bartending.foundations.step.FermentingBrewerStep;
 import ml.pluto7073.bartending.foundations.util.BrewingUtil;
 import ml.pluto7073.bartending.foundations.alcohol.AlcDisplayType;
 import ml.pluto7073.bartending.foundations.alcohol.AlcoholHandler;
@@ -103,6 +104,9 @@ public class PourableBottleItem extends Item {
         int amount = BrewingUtil.getStandardAlcohol(drink);
         amount += BrewingUtil.getAlcoholDeviation(stack);
         if (isAdvanced.isAdvanced() || isAdvanced.isCreative()) AlcoholHandler.INSTANCE.appendTooltip(tooltip, amount, stack, AlcDisplayType.PROOF);
+        if (stack.getOrCreateTag().contains("ExtraFermentingData")) {
+            FermentingBrewerStep.appendInProgressText(stack.getOrCreateTagElement("ExtraFermentingData"), tooltip);
+        }
         super.appendHoverText(stack, level, tooltip, isAdvanced);
     }
 }